moved cli/core/pkg/command/core files to cli/core/pkg/command and delete core directory directory Signed-off-by: Prem Kumar Kalle <pkalle@vmware.com>
