+++ cli/core/pkg/command/plugin_manager_test.go
@@ -1,7 +1,7 @@
 // Copyright 2022 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package core
+package command
 
 import (
 	"testing"
