+++ cli/core/pkg/command/root.go
@@ -1,7 +1,7 @@
 // Copyright 2021 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package core
+package command
 
 import (
 	"fmt"
