+++ cli/core/pkg/command/templates/README.md.tmpl
