+++ cli/core/pkg/command/templates/core.go
