+++ cli/core/pkg/command/version_test.go
@@ -1,7 +1,7 @@
 // Copyright 2021 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package core
+package command
 
 import (
 	"io"
